+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- **[FIX]** Fix falsy values in `ItemsList`
+- **[FIX]** Make input types in `TopBar` more flexible
 [...]
 
 # v45.0.1 (18/01/2021)